+++ frontend/playwright/ui/specs/tokens/crud.spec.js
@@ -8,6 +8,7 @@ import {
   testTokenCreationFlow,
   unfoldTokenType,
   createToken,
+  createSet,
 } from "./helpers";
 
 test.beforeEach(async ({ page }) => {
@@ -1135,7 +1136,7 @@ test.describe("Tokens - creation", () => {
     ).toBeEnabled();
   });
 
-  test("User cant submit empty typography token or reference", async ({
+  test("User can't submit empty typography token or reference", async ({
     page,
   }) => {
     const { tokensUpdateCreateModal, tokenThemesSetsSidebar, tokensSidebar } =
@@ -1693,7 +1694,7 @@ test.describe("Tokens - creation", () => {
     await expect(tokensSidebar.getByLabel("primary")).toBeEnabled();
   });
 
-  test("User cant create regular token with value missing", async ({
+  test("User can't create regular token with value missing", async ({
     page,
   }) => {
     const { tokensUpdateCreateModal } = await setupEmptyTokensFileRender(page);
@@ -1818,7 +1819,7 @@ test("User creates grouped color token", async ({ page }) => {
   await expect(tokensSidebar.getByLabel("primary")).toBeEnabled();
 });
 
-test("User cant create regular token with value missing", async ({ page }) => {
+test("User can't create regular token with value missing", async ({ page }) => {
   const { tokensUpdateCreateModal } = await setupEmptyTokensFileRender(page);
 
   const tokensTabPanel = page.getByRole("tabpanel", { name: "tokens" });
@@ -1897,6 +1898,339 @@ test("User disables the current set but token still have resolved values shown i
     "title",
     /Resolved value: #ff0000/,
   );
+});
+
+test.describe("User can't create groups that clash with token names", () => {
+  const createBadToken = async (page, type, name, textFieldName, value) => {
+    const tokensTabPanel = page.getByRole("tabpanel", { name: "tokens" });
+
+    const { tokensUpdateCreateModal } = await setupTokensFileRender(page, {
+      flags: ["enable-token-shadow"],
+    });
+
+    // Add a token of the given type
+    await tokensTabPanel
+      .getByRole("button", { name: `Add Token: ${type}` })
+      .click();
+    await expect(tokensUpdateCreateModal).toBeVisible();
+
+    // Fill the bad name
+    const nameField = tokensUpdateCreateModal.getByLabel("Name");
+    await nameField.fill(name);
+
+    // Fill the value
+    const valueField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: textFieldName,
+    });
+    await valueField.fill(value);
+
+    // Check that the value has an error
+    const errorNode =
+      tokensUpdateCreateModal.getByText(`Group name of ${name} conflicts with a token of the same name in another active set.`);
+
+    await expect(errorNode).toBeVisible();
+
+    // Check that the form cannot be saved
+    const submitButton = tokensUpdateCreateModal.getByRole("button", {
+      name: "Save",
+    });
+    await expect(submitButton).toBeDisabled();
+  };
+
+  test("User can't create Border Radius token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Border Radius", "rad1.bad", "Value", "10");
+  });
+
+  test("User can't create Color token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Color", "col1.bad", "Value", "red");
+  });
+
+  test("User can't create Dimensions token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Dimensions", "dim1.bad", "Value", "100");
+  });
+
+  test("User can't create Font Size token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Font Size", "fsiz1.bad", "Value", "16");
+  });
+
+  test("User can't create Font Weight token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Font Weight", "wei1.bad", "Value", "400");
+  });
+
+  test("User can't create Letter Spacing token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Letter Spacing", "lspa1.bad", "Value", "1");
+  });
+
+  test("User can't create Number token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Number", "num1.bad", "Value", "10");
+  });
+
+  test("User can't create Rotation token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Rotation", "rot1.bad", "Value", "90");
+  });
+
+  test("User can't create Sizing token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Sizing", "siz1.bad", "Value", "100");
+  });
+
+  test("User can't create Spacing token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Spacing", "spa1.bad", "Value", "10");
+  });
+
+  test("User can't create Stroke Width token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Stroke Width", "str1.bad", "Value", "2");
+  });
+
+  test("User can't create Text Case token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Text Case", "cas1.bad", "Value", "uppercase");
+  });
+
+test("User can't create Text Decoration token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar, tokenContextMenuForToken } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    await createBadToken(page, "Text Decoration", "dec1.bad", "Value", "strike-through");
+  });
+
+  test("User can't create Typography token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    const tokensTabPanel = page.getByRole("tabpanel", { name: "tokens" });
+
+    const { tokensUpdateCreateModal } = await setupTokensFileRender(page, {
+      flags: ["enable-token-shadow"],
+    });
+
+    await tokensTabPanel
+      .getByRole("button", { name: `Add Token: Typography` })
+      .click();
+    await expect(tokensUpdateCreateModal).toBeVisible();
+
+    const nameField = tokensUpdateCreateModal.getByLabel("Name");
+    await nameField.fill("typ1.bad");
+
+    const fontFamilyField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Font family",
+    });
+    await fontFamilyField.fill("Arial");
+
+    const fontSizeField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Font size",
+    });
+    await fontSizeField.fill("16");
+
+    const fontWeightField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Font weight",
+    });
+    await fontWeightField.fill("400");
+
+    const lineHeightField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Line height",
+    });
+    await lineHeightField.fill("1.5");
+
+    const letterSpacingField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Letter spacing",
+    });
+    await letterSpacingField.fill("0");
+
+    const textCaseField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Text case",
+    });
+    await textCaseField.fill("none");
+
+    const textDecorationField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Text decoration",
+    });
+    await textDecorationField.fill("none");
+
+    const submitButton = tokensUpdateCreateModal.getByRole("button", {
+      name: "Save",
+    });
+
+    const errorNode = tokensUpdateCreateModal.getByText("Group name of typ1.bad conflicts with a token of the same name in another active set.");
+    await expect(errorNode).toHaveCount(1);
+    await expect(submitButton).toBeDisabled();
+  });
+
+  test("User can't create Shadow token with group name that clashes with existing token", async ({ page }) => {
+    const { tokenThemesSetsSidebar, tokensSidebar } =
+      await setupTokensFileRender(page, {
+        file: "workspace/get-file-tokens-all-types.json"
+      });
+
+    await expect(tokensSidebar).toBeVisible();
+
+    await createSet(tokenThemesSetsSidebar, "Second set");
+
+    const tokensTabPanel = page.getByRole("tabpanel", { name: "tokens" });
+
+    const { tokensUpdateCreateModal } = await setupTokensFileRender(page, {
+      flags: ["enable-token-shadow"],
+    });
+
+    await tokensTabPanel
+      .getByRole("button", { name: `Add Token: Shadow` })
+      .click();
+    await expect(tokensUpdateCreateModal).toBeVisible();
+
+    const nameField = tokensUpdateCreateModal.getByLabel("Name");
+    await nameField.fill("sha1.bad");
+
+    const colorField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Color",
+    });
+    await colorField.fill("red");
+
+    const offsetXField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "X",
+    });
+    await offsetXField.fill("7");
+
+    const offsetYField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Y",
+    });
+    await offsetYField.fill("8");
+
+    const blurField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Blur",
+    });
+    await blurField.fill("5");
+
+    const spreadField = tokensUpdateCreateModal.getByRole("textbox", {
+      name: "Spread",
+    });
+    await spreadField.fill("1");
+
+    const submitButton = tokensUpdateCreateModal.getByRole("button", {
+      name: "Save",
+    });
+
+    const errorNode = tokensUpdateCreateModal.getByText("Group name of sha1.bad conflicts with a token of the same name in another active set.");
+    await expect(errorNode).toHaveCount(1);
+    await expect(submitButton).toBeDisabled();
+  });
 });
 
 test.describe("Tokens tab - edition", () => {
@@ -2122,7 +2456,7 @@ test.describe("Tokens tab - delete", () => {
   });
 });
 
-test("BUG: 1425 Token is not highlighted in red when value references a token in a disabled set", async ({
+test("BUG: 14262 Token pill must be highlighted when value references a token in a disabled set", async ({
   page,
 }) => {
   const { tokensSidebar, tokenContextMenuForToken, tokenThemesSetsSidebar } =
@@ -2140,15 +2474,47 @@ test("BUG: 1425 Token is not highlighted in red when value references a token in
     "textbox",
     "{base-radius}",
   );
-  
+
   const refTokenPill = tokensSidebar.getByRole("button", {
     name: "ref-base",
   });
 
   await expect(refTokenPill).toBeVisible();
 
-  const CoreSetCheckbox = tokenThemesSetsSidebar.getByRole('button', { name: 'core' }).getByRole('checkbox'); 
+  const CoreSetCheckbox = tokenThemesSetsSidebar
+    .getByRole("button", { name: "core" })
+    .getByRole("checkbox");
   await CoreSetCheckbox.click();
-  const brokenTokenPill = tokensSidebar.getByRole('button', { name: 'Missing reference ref-base' });
-  await expect(brokenTokenPill).toBeVisible();
+
+  // Pill is not highlighted if both tokens are on the same disabled set
+  const brokenTokenPill = tokensSidebar.getByRole("button", {
+    name: "Missing reference ref-base",
+  });
+  await expect(brokenTokenPill).not.toBeVisible();
+  await createSet(tokenThemesSetsSidebar, "New set");
+  await tokenThemesSetsSidebar.getByRole("button", { name: "New set" }).click();
+
+  await tokenThemesSetsSidebar
+    .getByRole("button", { name: "New set" })
+    .getByRole("checkbox")
+    .click();
+  await createToken(page, "Border radius", "new-ref", "Value", "textbox", "{base-radius}");
+
+  // Pill is highlighted if the referenced token is in a different disabled set than the token with the reference
+  const newBrokenTokenPill = tokensSidebar.getByRole("button", {
+    name: "Missing reference new-ref",
+  });
+  await expect(newBrokenTokenPill).toBeVisible();
+  await tokenThemesSetsSidebar
+    .getByRole("button", { name: "core" })
+    .getByRole("checkbox")
+    .click();
+
+  // When the disabled set is activated again, pill is not highlighted anymore
+  await expect(
+    tokenThemesSetsSidebar
+      .getByRole("button", { name: "core" })
+      .getByRole("checkbox"),
+  ).toBeChecked();
+  await expect(newBrokenTokenPill).not.toBeVisible();
 });